refactor methods for checks questions
